+++ CHANGELOG.md
@@ -22,6 +22,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Bug Fixes
 
+* Fix incremental installation with plugins that include arguments with different ordering.  
+  [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
+  [#9170](https://github.com/CocoaPods/CocoaPods/pull/9170)
+
 * Move custom `Copy Headers` script phase for header mappings before `Compile Sources`.  
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#9131](https://github.com/CocoaPods/CocoaPods/pull/9131)